LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/scroll-restoration-order-expected.txt


FAIL Traversing history should restore scroll position after dispatching popstate and before dispatching hashchange assert_array_approx_equals: scroll is restored before hashchange for #2 property 0, expected 100 +/- undefined, expected 100 but got 555
